requirements.txt
antiorm==1.2.1
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
db==0.1.1
et-xmlfile==1.1.0
idna==3.4
openpyxl==3.0.10
pysqlite3-binary==0.4.7.post7
requests==2.28.1
soupsieve==2.3.2.post1
urllib3==1.26.12
